api/pyproject.toml
[tool.poetry]
name = "fastapi-socialnetworking"
version = "0.1.0"
description = ""
authors = ["Svetlana Vlad <sv.vlad@diogen.live>"]
readme = "README.md"
packages = [{include = "fastapi_socialnetworking"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.18"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
aiohttp = "^3.8.4"
passlib = "^1.7.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
